Move click handler from Map component to caller
The Map component should ideally just render the map and pass events to
handlers provided by the caller. This way the logic what should be done
with a click event is defined wherever that map is rendered. In the end
this makes it possible to separate functionality between admin and user
sides.